application/requirements.txt
requests
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-requests
opentelemetry-exporter-opencensus